// apps/web/lib/gen/model/businessCreatePayload.ts
/**
 * Generated by orval v8.0.0-rc.1 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */

export interface BusinessCreatePayload {
  name: string;
  /** @nullable */
  description?: string | null;
  contactNumber: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  gcashQrUrl?: string | null;
}